requirements.txt
aioesphomeapi
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
influxdb-client